requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-google-genai
faiss-cpu
pymupdf
python-dotenv